scrapers/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
libsql-experimental>=0.0.34
python-dotenv>=1.0.0
